Finished version 0.0 of the b2s_clf library
