+++ b2s_clf/dataset_transformer/signal_compressor.py
