+++ b2s_clf/ensemble/ensemble.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pandas as pd
 
-from vgc_clf.sampler.sampler import Sampler
+from b2s_clf.sampler.sampler import Sampler
 
 
 class Ensemble:
